% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_points.R
\name{draw_points}
\alias{draw_points}
\title{run shiny app : making points by click on the map}
\usage{
draw_points()
}
\value{

}
\description{
run shiny app : making points by click on the map
}
\examples{
draw_points()
}
