% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_lines.R
\name{draw_lines}
\alias{draw_lines}
\title{run shiny app : making lines by click on the map}
\usage{
draw_lines()
}
\value{

}
\description{
run shiny app : making lines by click on the map
}
\examples{
draw_lines()
}
