% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_lines.R
\name{draw_lines}
\alias{draw_lines}
\title{Making Lines by Clicks on the Map}
\usage{
draw_lines()
}
\value{
An object that represents the app. Printing the object or passing it to shiny::runApp() will run the app.
}
\description{
Making Lines by Clicks on the Map
}
\examples{
# Only run this example in interactive R sessions
# draw_lines()
}
