% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_points.R
\name{draw_points}
\alias{draw_points}
\title{Making Points by Clicks on the Map}
\usage{
draw_points()
}
\value{
An object that represents the app. Printing the object or passing it to shiny::runApp() will run the app.
}
\description{
Making Points by Clicks on the Map
}
\examples{
# Only run this example in interactive R sessions
if (interactive()) draw_points()
}
