% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_points.R
\name{draw_points}
\alias{draw_points}
\title{Making points by clicks on the map}
\usage{
draw_points()
}
\value{
An object that represents the app. Printing the object or passing it to shiny::runApp() will run the app.
}
\description{
This function run a Shiny application with map.
You can make your own geospatial point by click on the map.
If you click save button then your geospatial points will be saved as a RDS file.
}
\examples{
# Only run this example in interactive R sessions
if (interactive()) draw_points()
}
