% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_polygons.R
\name{draw_polygons}
\alias{draw_polygons}
\title{run shiny app : making polygons by click on the map}
\usage{
draw_polygons()
}
\value{

}
\description{
run shiny app : making polygons by click on the map
}
\examples{
draw_polygons()
}
