% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_polygons.R
\name{draw_polygons}
\alias{draw_polygons}
\title{Making polygons by clicks on the map}
\usage{
draw_polygons()
}
\value{
An object that represents the app. Printing the object or passing it to shiny::runApp() will run the app.
}
\description{
This function run a Shiny application with map.
You can shape your own geospatial polygon by clicks on the map and complete it with make button.
If you click save button then your completed geospatial polygons will be saved as a RDS file.
}
\examples{
# Only run this example in interactive R sessions
if (interactive()) draw_polygons()
}
