% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_polygons.R
\name{draw_polygons}
\alias{draw_polygons}
\title{Making Polygons by Clicks on the Map}
\usage{
draw_polygons()
}
\value{
An object that represents the app. Printing the object or passing it to shiny::runApp() will run the app.
}
\description{
Making Polygons by Clicks on the Map
}
\examples{
# Only run this example in interactive R sessions
if (interactive()) draw_polygons()
}
